# product-service/src/main/resources/product-api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/product-apis
  description: Generated server url
paths:
  /product-apis/product:
    get:
      tags:
      - product-controller
      operationId: getAllProduct
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
      - product-controller
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
  /product-apis/product/{productId}:
    get:
      tags:
      - product-controller
      operationId: getProductByProductId
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
      - product-controller
      operationId: deleteProductByProductId
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    patch:
      tags:
      - product-controller
      operationId: updateProductByProductId
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Product"
  /product-apis:
    get:
      tags:
      - product-controller
      operationId: test
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
          minLength: 1
        price:
          type: number
          format: double
      required:
      - name